requirements.txt
bcrypt==3.1.7
cffi==1.14.0
click==7.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-JWT-Extended==3.24.1
Flask-SQLAlchemy==2.4.3
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2==2.8.5
pycparser==2.20
PyJWT==1.7.1
PyPika==0.37.15
python-dateutil==2.8.1
python-dotenv==0.13.0
six==1.15.0
SQLAlchemy==1.3.18
Werkzeug==1.0.1
